Add files via upload
Problem: Suppose you are making an experiment in which you will get 2 number combinations. You are
given 12 ping pong balls with two different colors (yellow and white), each colored ball is
numbered 1 to 6. Create a flowchart in which you will release 2 numbered balls, 1 of each color,
randomly, 6 times maximum. You are also given 3 colored baskets (yellow, red and white). If the
number in each ball is the same, should go to red. If the product of two balls released is even, the
ball should go to white basket, if odd then to yellow.